use RuntimeReplaceable with StaticInvoke
