+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/variant/VariantExpressionSuite.scala
@@ -17,7 +17,7 @@
 
 package org.apache.spark.sql.catalyst.expressions.variant
 
-import org.apache.spark.{SparkException, SparkFunSuite, SparkRuntimeException}
+import org.apache.spark.{SparkFunSuite, SparkRuntimeException}
 import org.apache.spark.sql.catalyst.analysis.ResolveTimeZone
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.types.variant.VariantUtil._
@@ -32,141 +32,6 @@ class VariantExpressionSuite extends SparkFunSuite with ExpressionEvalHelper {
       val padding = List.fill(size - 1)(0.toByte)
       b :: padding
     }
-  }
-
-  test("parse_json") {
-    def check(json: String, expectedValue: Array[Byte], expectedMetadata: Array[Byte]): Unit = {
-      checkEvaluation(ParseJson(Literal(json)), new VariantVal(expectedValue, expectedMetadata))
-    }
-
-    // Dictionary size is `0` for value 0. An empty dictionary contains one offset `0` for the
-    // one-past-the-end position (i.e. the sum of all string lengths).
-    val emptyMetadata = Array[Byte](VERSION, 0, 0)
-    check("null", Array(primitiveHeader(NULL)), emptyMetadata)
-    check("true", Array(primitiveHeader(TRUE)), emptyMetadata)
-    check("false", Array(primitiveHeader(FALSE)), emptyMetadata)
-    check("1", Array(primitiveHeader(INT1), 1), emptyMetadata)
-    check("-1", Array(primitiveHeader(INT1), -1), emptyMetadata)
-    check("127", Array(primitiveHeader(INT1), 127), emptyMetadata)
-    check("128", Array(primitiveHeader(INT2), -128, 0), emptyMetadata)
-    check("-32768", Array(primitiveHeader(INT2), 0, -128), emptyMetadata)
-    check("-32769", Array(primitiveHeader(INT4), -1, 127, -1, -1), emptyMetadata)
-    check("2147483647", Array(primitiveHeader(INT4), -1, -1, -1, 127), emptyMetadata)
-    check("2147483648", Array(primitiveHeader(INT8), 0, 0, 0, -128, 0, 0, 0, 0), emptyMetadata)
-    check("9223372036854775807",
-      Array(primitiveHeader(INT8), -1, -1, -1, -1, -1, -1, -1, 127), emptyMetadata)
-    check("-9223372036854775808",
-      Array(primitiveHeader(INT8), 0, 0, 0, 0, 0, 0, 0, -128), emptyMetadata)
-    check("9223372036854775808",
-      Array(primitiveHeader(DECIMAL16), 0, 0, 0, 0, 0, 0, 0, 0, -128, 0, 0, 0, 0, 0, 0, 0, 0),
-      emptyMetadata)
-    check("1.0", Array(primitiveHeader(DECIMAL4), 1, 10, 0, 0, 0), emptyMetadata)
-    check("1.01", Array(primitiveHeader(DECIMAL4), 2, 101, 0, 0, 0), emptyMetadata)
-    check("99999.9999", Array(primitiveHeader(DECIMAL4), 4, -1, -55, -102, 59), emptyMetadata)
-    check("99999.99999",
-      Array(primitiveHeader(DECIMAL8), 5, -1, -29, 11, 84, 2, 0, 0, 0), emptyMetadata)
-    check("0.000000001", Array(primitiveHeader(DECIMAL4), 9, 1, 0, 0, 0), emptyMetadata)
-    check("0.0000000001",
-      Array(primitiveHeader(DECIMAL8), 10, 1, 0, 0, 0, 0, 0, 0, 0), emptyMetadata)
-    check("9" * 38,
-      Array[Byte](primitiveHeader(DECIMAL16), 0) ++ BigInt("9" * 38).toByteArray.reverse,
-      emptyMetadata)
-    check("1" + "0" * 38,
-      Array(primitiveHeader(DOUBLE)) ++
-        BigInt(java.lang.Double.doubleToLongBits(1E38)).toByteArray.reverse,
-      emptyMetadata)
-    check("\"\"", Array(shortStrHeader(0)), emptyMetadata)
-    check("\"abcd\"", Array(shortStrHeader(4), 'a', 'b', 'c', 'd'), emptyMetadata)
-    check("\"" + ("x" * 63) + "\"",
-      Array(shortStrHeader(63)) ++ Array.fill(63)('x'.toByte), emptyMetadata)
-    check("\"" + ("y" * 64) + "\"",
-      Array[Byte](primitiveHeader(LONG_STR), 64, 0, 0, 0) ++ Array.fill(64)('y'.toByte),
-      emptyMetadata)
-    check("{}", Array(objectHeader(false, 1, 1),
-      /* size */ 0,
-      /* offset list */ 0), emptyMetadata)
-    check("[]", Array(arrayHeader(false, 1),
-      /* size */ 0,
-      /* offset list */ 0), emptyMetadata)
-    check("""{"a": 1, "b": 2, "c": "3"}""", Array(objectHeader(false, 1, 1),
-      /* size */ 3,
-      /* id list */ 0, 1, 2,
-      /* offset list */ 0, 2, 4, 6,
-      /* field data */ primitiveHeader(INT1), 1, primitiveHeader(INT1), 2, shortStrHeader(1), '3'),
-      Array(VERSION, 3, 0, 1, 2, 3, 'a', 'b', 'c'))
-    check("""{"z": 1, "y": 2, "x": "3"}""", Array(objectHeader(false, 1, 1),
-      /* size */ 3,
-      /* id list */ 2, 1, 0,
-      /* offset list */ 4, 2, 0, 6,
-      /* field data */ primitiveHeader(INT1), 1, primitiveHeader(INT1), 2, shortStrHeader(1), '3'),
-      Array(VERSION, 3, 0, 1, 2, 3, 'z', 'y', 'x'))
-    check("""[null, true, {"false" : 0}]""", Array(arrayHeader(false, 1),
-      /* size */ 3,
-      /* offset list */ 0, 1, 2, 9,
-      /* element data */ primitiveHeader(NULL), primitiveHeader(TRUE), objectHeader(false, 1, 1),
-      /* size */ 1,
-      /* id list */ 0,
-      /* offset list */ 0, 2,
-      /* field data */ primitiveHeader(INT1), 0),
-      Array(VERSION, 1, 0, 5, 'f', 'a', 'l', 's', 'e'))
-  }
-
-  test("parse_json negative") {
-    for (json <- Seq("", "[", "+1", "1a", """{"a": 1, "b": 2, "a": "3"}""")) {
-      checkExceptionInExpression[SparkException](ParseJson(Literal(json)),
-        "Malformed records are detected in record parsing")
-    }
-    for (json <- Seq("\"" + "a" * (16 * 1024 * 1024) + "\"",
-      (0 to 4 * 1024 * 1024).mkString("[", ",", "]"))) {
-      checkExceptionInExpression[SparkRuntimeException](ParseJson(Literal(json)),
-        "Cannot build variant bigger than 16.0 MiB")
-    }
-  }
-
-  test("round-trip") {
-    def check(input: String, output: String = null): Unit = {
-      checkEvaluation(
-        StructsToJson(Map.empty, ParseJson(Literal(input))),
-        if (output != null) output else input
-      )
-    }
-
-    check("null")
-    check("true")
-    check("false")
-    check("-1")
-    check("1.0E10")
-    check("\"\"")
-    check("\"" + ("a" * 63) + "\"")
-    check("\"" + ("b" * 64) + "\"")
-    // scalastyle:off nonascii
-    check("\"" + ("你好，世界" * 20) + "\"")
-    // scalastyle:on nonascii
-    check("[]")
-    check("{}")
-    // scalastyle:off nonascii
-    check(
-      "[null, true,   false,-1, 1e10, \"\\uD83D\\uDE05\", [ ], { } ]",
-      "[null,true,false,-1,1.0E10,\"😅\",[],{}]"
-    )
-    // scalastyle:on nonascii
-    check("[0.0, 1.00, 1.10, 1.23]", "[0,1,1.1,1.23]")
-  }
-
-  test("to_json with nested variant") {
-    checkEvaluation(
-      StructsToJson(Map.empty, CreateArray(Seq(ParseJson(Literal("{}")),
-        ParseJson(Literal("\"\"")),
-        ParseJson(Literal("[1, 2, 3]"))))),
-      "[{},\"\",[1,2,3]]"
-    )
-    checkEvaluation(
-      StructsToJson(Map.empty, CreateNamedStruct(Seq(
-        Literal("a"), ParseJson(Literal("""{ "x": 1, "y": null, "z": "str" }""")),
-        Literal("b"), ParseJson(Literal("[[]]")),
-        Literal("c"), ParseJson(Literal("false"))))),
-      """{"a":{"x":1,"y":null,"z":"str"},"b":[[]],"c":false}"""
-    )
   }
 
   test("to_json malformed") {
